Avoid possible test failures in t/02_function.t, t/03_script and t/04_par.t
From MSDN's documentation of DeleteFile(), "The DeleteFile function marks a
file for deletion on close. Therefore, the file deletion does not occur
until the last handle to the file is closed. Subsequent calls to CreateFile
to open the file fail with ERROR_ACCESS_DENIED."

So if another process, typically anti-virus software or programs like
TSVNCache.exe or TGitCache.exe, has a handle open on a file being unlink()ed
then it does not get deleted straight away and an attempt to open the same
file again straight afterwards will fail because of the pending delete.

Rather than playing games with either detecting the pending delete (using
GetFileInformationByHandleEx()) or simply sleeping and retrying the open
anyway, it is simpler in these cases to just use a different filename for
each test.