Tolerate spaces in paths in Makefile.PL
Tolerate spaces in the path to openssl, which is common in Windows,
by quoting $bin_file in the string passed to backticks for execution.

This avoids this result when openssl is in a directory with spaces:

'C:\Program' is not recognized as an internal or external command,
operable program or batch file.